+++ web/index.php
@@ -1,25 +1,5 @@
 <?php
 
-/**
- * Костыль для переадресаций index.php, index.html на соотв. роуты, пример: '/index.php/news' -> '/news'
- * Данное решение было сделано для облегчения настройки веб-сервера и выполнения SEO-требований
- */
-function checkPoint()
-{
-    if (strpos($_SERVER['REQUEST_URI'], 'index.php') || strpos($_SERVER['REQUEST_URI'], 'index.html')) {
-        header('HTTP/1.0 301 Moved Permanently');
-        $replaced_url = str_replace(
-            ['index.php/', 'index.php', 'index.html'],
-            ['', '', ''],
-            str_replace('?', '', $_SERVER['REQUEST_URI'])
-        );
-        header('Location: http://' . $_SERVER['HTTP_HOST'] . $replaced_url);
-        exit;
-    }
-}
-
-checkPoint();
-
 chdir(dirname(__DIR__));
 
 define('ROOT', __DIR__);
@@ -27,8 +7,7 @@ define('HOST_HASH', substr(md5($_SERVER['HTTP_HOST']), 0, 12));
 
 defined('APPLICATION_ENV') || define('APPLICATION_ENV', (getenv('APPLICATION_ENV') ? getenv('APPLICATION_ENV') : 'production'));
 
-define('APPLICATION_PATH', (APPLICATION_ENV == 'development') ? __DIR__ . '/../app' : __DIR__ . '/../app');
-define('PUBLIC_PATH', __DIR__);
+define('APPLICATION_PATH', __DIR__ . '/../app');
 
 require_once APPLICATION_PATH . '/Bootstrap.php';
 $bootstrap = new YonaCMS\Bootstrap();